pyproject.toml
[project]
name = "australia-epa"
version = "0.1.0"
description = "dev env for API calls"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]
